@@ -11,80 +10,81 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
-    "@ctrl/ngx-headroom": "^2.3.2",
-    "@ng-bootstrap/ng-bootstrap": "^5.1.4",
-    "@ngrx/effects": "^8.5.2",
-    "@ngrx/store": "^8.5.2",
-    "@ngrx/store-devtools": "^8.5.2",
+    "@angular/animations": "^17.0.8",
+    "@angular/cdk": "^17.0.4",
+    "@angular/common": "^17.0.8",
+    "@angular/compiler": "^17.0.8",
+    "@angular/core": "^17.0.8",
+    "@angular/forms": "^17.0.8",
+    "@angular/material": "^17.0.4",
+    "@angular/platform-browser": "^17.0.8",
+    "@angular/platform-browser-dynamic": "^17.0.8",
+    "@angular/router": "^17.0.8",
+    "@ctrl/ngx-headroom": "^5.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+    "@ngrx/effects": "^17.0.1",
+    "@ngrx/store": "^17.0.1",
+    "@ngrx/store-devtools": "^17.0.1",
+    "@schematics/angular": "^17.0.9",
     "angular-timer": "^1.3.5",
     "angularx-social-login": "^1.2.7",
-    "body-parser": "^1.19.0",
-    "bootstrap": "^4.4.1",
-    "connect-mongo": "^3.2.0",
-    "cookie-parser": "^1.4.4",
+    "body-parser": "^1.20.2",
+    "bootstrap": "^5.3.2",
+    "connect-mongo": "^5.1.0",
+    "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "crypto": "^1.0.1",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "express-fileupload": "^1.1.6",
-    "express-form-data": "^2.0.10",
-    "express-session": "^1.17.0",
-    "express-validator": "^6.3.0",
-    "hamburgers": "^1.1.3",
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "express-fileupload": "^1.4.3",
+    "express-form-data": "^2.0.23",
+    "express-session": "^1.17.3",
+    "express-validator": "^7.0.1",
+    "hamburgers": "^1.2.1",
     "hammerjs": "^2.0.8",
-    "headroom.js": "^0.9.4",
-    "jquery": "^3.4.1",
-    "material-community-components": "^5.0.1",
-    "moment": "^2.24.0",
-    "mongoose": "^5.8.1",
+    "headroom.js": "^0.12.0",
+    "jquery": "^3.7.1",
+    "material-community-components": "^13.0.0",
+    "moment": "^2.30.1",
+    "mongoose": "^8.0.3",
     "multer": "^1.4.2",
-    "ngx-countdown": "^3.2.0",
-    "ngx-doc-viewer": "^0.1.25",
-    "ngx-summernote": "^0.6.5",
-    "node-schedule": "^1.3.2",
-    "nodemailer": "^6.4.2",
+    "ngx-countdown": "^17.1.1",
+    "ngx-doc-viewer": "^15.0.1",
+    "ngx-summernote": "^1.0.0",
+    "node-schedule": "^2.1.1",
+    "nodemailer": "^6.9.8",
     "open-iconic": "^1.1.1",
-    "passport": "^0.4.1",
+    "passport": "^0.7.0",
     "passport-facebook": "^3.0.0",
     "passport-google-oauth": "^2.0.0",
-    "popper.js": "^1.16.0",
-    "rxjs": "^6.5.3",
-    "save": "^2.4.0",
-    "summernote": "^0.8.12",
-    "tslib": "^1.10.0",
-    "validator": "^11.1.0",
-    "zone.js": "~0.9.1"
+    "popper.js": "^1.16.1",
+    "rxjs": "^7.8.1",
+    "save": "^2.9.0",
+    "summernote": "^0.8.20",
+    "tslib": "^2.6.2",
+    "validator": "^13.11.0",
+    "zone.js": "^0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.802.2",
-    "@angular/cli": "^8.3.20",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
-    "@ckeditor/ckeditor5-alignment": "^11.2.0",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "^2.0.8",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.2.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "@angular-devkit/build-angular": "^17.0.9",
+    "@angular/cli": "^17.0.9",
+    "@angular/compiler-cli": "^17.0.8",
+    "@angular/language-service": "^17.0.8",
+    "@ckeditor/ckeditor5-alignment": "^40.2.0",
+    "@types/jasmine": "~5.1.4",
+    "@types/jasminewd2": "^2.0.13",
+    "@types/node": "~20.10.6",
+    "codelyzer": "^0.0.28",
+    "jasmine-core": "~5.1.1",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.4.2",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~10.9.2",
+    "tslint": "~5.20.1",
+    "typescript": ">=5.2.0 <5.3.0"
   }
 }
