@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
-import { MatDrawer } from '@angular/material';
-import { BehaviorSubject } from 'rxjs';
+import { MatDrawer } from '@angular/material/sidenav';
 
 @Injectable({
   providedIn: 'root'
